requirements.txt
amqp==5.0.2
appdirs==1.4.4
asgiref==3.2.10
attrs==20.3.0
autopep8==1.5.4
bcrypt==3.2.0
billiard==3.6.3.0
Bootstrap-Flask==1.5.1
cached-property==1.5.2
celery==5.0.5
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
crcmod==1.7
cryptography==3.1.1
distlib==0.3.1
distro==1.5.0
Django==3.1.1
docker==4.3.1
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
filelock==3.0.12
Flask==1.1.2
Flask-Caching==1.9.0
Flask-SQLAlchemy==2.4.4
future==0.18.2
googlemaps==4.4.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
kombu==5.0.2
line-bot-sdk==1.18.0
MarkupSafe==1.1.1
paramiko==2.7.2
prompt-toolkit==3.0.8
pycodestyle==2.6.0
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.22
sqlparse==0.3.1
texttable==1.6.3
toml==0.10.1
urllib3==1.25.10
vine==5.0.0
virtualenv==20.0.34
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
zipp==3.1.0
